+++ attributes/appstream.rb
@@ -1,7 +1,8 @@
 default['yum']['appstream']['repositoryid'] = 'appstream'
-default['yum']['appstream']['description'] = "CentOS-#{node['platform_version'].to_i} - AppStream"
+default['yum']['appstream']['description'] = 'CentOS-$releasever - AppStream'
 default['yum']['appstream']['enabled'] = true
 default['yum']['appstream']['make_cache'] = true
 default['yum']['appstream']['managed'] = true
-default['yum']['appstream']['mirrorlist'] = "http://mirrorlist.centos.org/?release=#{node['platform_version'].to_i}&arch=$basearch&repo=AppStream&infra=$infra"
+default['yum']['appstream']['mirrorlist'] =
+  'http://mirrorlist.centos.org/?release=$releasever&arch=$basearch&repo=AppStream'
 default['yum']['appstream']['gpgkey'] = 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial'