+++ attributes/fasttrack.rb
@@ -1,6 +1,6 @@
 default['yum']['fasttrack']['repositoryid'] = 'fasttrack'
-default['yum']['fasttrack']['description'] = "CentOS-#{node['platform_version'].to_i} - fasttrack"
-default['yum']['fasttrack']['mirrorlist'] = "http://mirrorlist.centos.org/?release=#{node['platform_version'].to_i}&arch=$basearch&repo=fasttrack&infra=$infra"
+default['yum']['fasttrack']['description'] = 'CentOS-$releasever - fasttrack'
+default['yum']['fasttrack']['mirrorlist'] = 'http://mirrorlist.centos.org/?release=$releasever&arch=$basearch&repo=fasttrack'
 default['yum']['fasttrack']['enabled'] = false
 default['yum']['fasttrack']['make_cache'] = true
 default['yum']['fasttrack']['managed'] = false
@@ -9,5 +9,5 @@ default['yum']['fasttrack']['gpgkey'] =
   if node['platform_version'].to_i >= 8
     'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial'
   else
-    "file:///etc/pki/rpm-gpg/RPM-GPG-KEY-CentOS-#{node['platform_version'].to_i}"
+    'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-CentOS-$releasever'
   end