+++ attributes/powertools.rb
@@ -3,5 +3,5 @@ default['yum']['powertools']['description'] = "CentOS-#{node['platform_version']
 default['yum']['powertools']['enabled'] = false
 default['yum']['powertools']['make_cache'] = true
 default['yum']['powertools']['managed'] = false
-default['yum']['powertools']['mirrorlist'] = "http://mirrorlist.centos.org/?release=#{node['platform_version'].to_i}&arch=$basearch&repo=PowerTools&infra=$infra"
+default['yum']['powertools']['mirrorlist'] = "http://mirrorlist.centos.org/?release=#{node['platform_version'].to_i}&arch=$basearch&repo=PowerTools"
 default['yum']['powertools']['gpgkey'] = 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-centosofficial'